Move ImagerSharp to Examples
