+++ src/Examples/Examples.csproj
@@ -25,6 +25,7 @@
   <ItemGroup>
     <PackageReference Include="SharpZipLib" Version="1.3.3" />
     <PackageReference Include="SkiaSharp" Version="2.80.2" />
+    <PackageReference Include="SixLabors.ImageSharp" Version="2.1.1" />
     <PackageReference Include="System.Memory" Version="4.5.4" />
   </ItemGroup>
