+++ src/TorchSharp/TorchSharp.csproj
@@ -22,7 +22,6 @@
 
   <ItemGroup>
     <PackageReference Include="SharpZipLib" Version="1.3.3" />
-    <PackageReference Include="SixLabors.ImageSharp" Version="2.1.1" />
     <PackageReference Include="System.Memory" Version="4.5.4" />
   </ItemGroup>
